requirements.txt
ipywidgets==8.1.5
ipykernel==6.29.5
bitsandbytes==0.45.2
accelerate==1.4.0
xformers==0.0.29
peft==0.14.0
trl==0.15.1
triton==3.1.0
unsloth-zoo==2025.2.7
unsloth==2025.2.15
sentence-transformers==3.4.1
faiss-cpu==1.10.0
spacy==3.8.4
more-itertools==10.6.0
transformers==4.49.0
qwen-vl-utils[decord]==0.0.10
wandb==0.19.8
flash-attn==2.7.4.post1
openai==1.75.0
lm-eval==0.4.8
tqdm-joblib==0.0.4